Extract observer for active account into dedicated action
It is only used once so far, so there's no point in running it for every app instance
